org-daemon: update internal structure of open dates
Previously, the EDIT button of an open date did not jump at the
correct location within emacs if the org file was edited and the date
moved. This is now fixed; the internal date object is correctly
updated during the regular update checks, or if the Update button was
clicked.

To implement this, the Emacs::Org::Daemon::Date class has a new copy()
method.