reports in sync directory are treated as "old" ones
This is what the user expects: in the "N+M" tag, the N is the number
of already seen reports, and this includes also the reports in the
"sync" directory. And M is only the number still in the "new"
directory.

This works also better if ctr_good_or_invalid is called while
send_tr_reports is still running.