simpler detection of missing modules (using module_exists)
This means CPAN::Reporter::PrereqCheck and the exception for
Term::ReadLine::Perl is not necessary anymore. This also means that
the check is more lightweight.

toolchain_modules_installed_check() now returns a list of missing modules,
so cpan_smoke_modules will just install these, and won't do any downgrades
(which can happen in the case of Expect).